requirements.txt
kafka-python
python-dotenv
faker
boto3